third commit

+++ Assignment/question-1.c
@@ -2,19 +2,19 @@
 #include<math.h>
 int main()
 {
-    int p,t,r;
+    float p,t,r;
     printf("Enter the principal amount : ");
-    scanf("%d",&p);
+    scanf("%f",&p);
     printf("Enter the Rate of interest : ");
-    scanf("%d",&r);
+    scanf("%f",&r);
     printf("Enter the time : ");
-    scanf("%d",&t);
+    scanf("%f",&t);
     float si=(p*t*r)/100;
     printf("The simple interest is : %f\n",si);
-    double Amount = p *  
+    float Amount = p *  
                   ((pow((1 + r / 100),t)));
-    double ci=Amount-p;
-    printf("The compound interest is : %lf\n",ci);
+    float ci=Amount-p;
+    printf("The compound interest is : %f\n",ci);
     
     return 0;
 }
